+++ src/controller_device_recv.c
@@ -264,7 +264,7 @@ device_state_recv_config(clixon_handle h,
             clicon_err(OE_UNIX, errno, "cbuf_new");
             goto done;
         }
-        cprintf(cberr, "YANG binding failed at mountpoint:");
+        cprintf(cberr, "YANG bind failed at mountpoint:");
         if ((x=xpath_first(xerr, NULL, "//error-message"))!=NULL)
             cprintf(cberr, "%s", xml_body(x));
         if (device_close_connection(dh, "%s", cbuf_get(cberr)) < 0)